loads data into battle screen and renders animated gifs
