--- startingdummy.js
@@ -1,30 +1,0 @@
-
-export const dummyData =
-{
-  instructorDexID: 1,
-  instructorName: 'Eric Katz',
-  description: 'Eric Katz AKA Prof. His insistence on wearing only white makes him vulnerable to coffee based attacks',
-  maxAttack: 50,
-  maxDefense: 20,
-  maxHP: 100,
-  moveSet: [
-    {
-      move: 'Use Vim instead of VSCode',
-      attack: 20,
-      type: 'Built Different'
-    },
-    {
-      move: 'Rub hair oil into beard',
-      attack: 15,
-      type: 'Built Different'
-    },
-    {
-      move: 'Reassign HELP ticket',
-      attack: 20,
-      type: 'Built Different'
-    }
-  ],
-  type: 'Built Different',
-  image: 'www.yahoo.com'
-}
-
